Fix logarithmic slowdown in ts_node_copy_child_nodes (#66)
This loop was slow for two reasons: firstly, we were using
`ts_node_child` to access the children of a given node, which entails
a logarithmic slowdown as the number of children exists, since these
nodes are stored in a balanced binary tree. Switching to the
`TSCursor` API means that we can traverse the node's children in
constant per-step time.

This also retools ts_node_elaborate to operate solely on pointer
values rather than incur the effort of allocating and returning a
struct value and then copying it into a pointer location.

On an array with 175000 elements, this reduces processing time from 12
minutes to less than 3 seconds.