Inline and avoid partial application in evalStruct.
Since the `Storable` instance for `Node` is implemented in terms of
the `evalStruct` function, it behooves us to avoid partial application
and function composition in its definition, as partially-applied
functions are slower than fully-applied ones.

When compiling with -O0, this produces _drastically_ faster parses on
large files: on a 5MB Go file parses went from 17sec to 7sec. The
improvement is less noticeable on when compiling with -O2, but is
still present (~0.5sec).